Fix the time lag during releases between index update to package update
During the release proess, debian/rpm index files on pkg.jenkins-ci.org
gets updated before actual package files appear in the mirrors.

This causes 404 until the mirroring kicks in, which can be several hours
in the bad case.

The release process is now modified to update index in
/var/www/pkg.jenkins-ci.org.staging. This change synchronizes the
delivery of bits to mirrors (at least to the fallback mirror) and
the updates to the repo index